test(unit): add global afterEach to cleanup antd components
