+++ src/components/designer/lightning/actions/PayInvoiceModal.spec.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { fireEvent, wait } from '@testing-library/dom';
-import { Modal, notification } from 'antd';
 import { Status } from 'shared/types';
 import { initChartFromNetwork } from 'utils/chart';
 import {
@@ -38,11 +37,6 @@ describe('PayInvoiceModal', () => {
       network,
     };
   };
-
-  afterEach(async () => {
-    Modal.destroyAll();
-    notification.destroy();
-  });
 
   it('should render labels', async () => {
     const { getByText } = await renderComponent();
